pyproject.toml
[build-system]
requires = ["setuptools>=42.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "domainforge"
version = "0.1.0"
description = "Generate full-stack applications from domain models"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "GPL-3.0"}
authors = [
    {name = "DomainForge Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "lark>=1.1.5",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "jinja2>=3.0.0",
    "click>=8.0.0",
    "fastapi>=0.100.0",
    "sqlalchemy>=2.0.0",
    "httpx>=0.24.0",
    "pytest-asyncio>=0.21.0",
    "aiosqlite>=0.19.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "pytest-mock>=3.0.0",
    "pytest-bdd>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=0.910.0",
    "pre-commit>=2.15.0"
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings>=0.24.0",
    "mkdocstrings-python>=1.7.0",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.0"
]

[project.scripts]
domainforge = "domainforge.cli:main"

[tool.setuptools]
package-dir = {"" = "domainforge"}
py-modules = ["domainforge"]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
ignore = []
select = ["E", "F", "D", "N", "S", "B", "C90"]

[tool.ruff.lint.isort]
known-first-party = ["domainforge"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
disallow_untyped_calls = true
disallow_untyped_globals = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "session"
